D2 1984

+++ SW Expert Academy/D2/1984/src/Solution.java
@@ -15,7 +15,7 @@ public class Solution {
 			Arrays.sort(arr);
 			
 			sum -= arr[0]+arr[9];
-			double ans = sum/8;
+			double ans = sum/(double)8;
 			System.out.println("#"+t+" "+Math.round(ans));
 		}
 	}
